Created Tunable int type for health and other int type uses.

+++ Entities/Players/Player.cs
@@ -38,7 +38,7 @@ public class Player : IPlayer
     private float breakTimer;
 
     //Stats
-    private readonly int maxHealth = (int)Tunables.PlayerMaxHealth.Value;
+    private readonly int maxHealth = Tunables.PlayerMaxHealth.Value;
     public int MaxHealth => maxHealth;
     private int health;
     public int Health => health;
